System.Process.Typed

Description

Please see the README.md file for examples of using this API.

Types

data ProcessConfig stdin stdout stderr

An abstract configuration for a process, which can then be launched into an actual running Process. Takes three type parameters, providing the types of standard input, standard output, and standard error, respectively.

There are three ways to construct a value of this type:

With the proc smart constructor, which takes a command name and a list of arguments.

With the shell smart constructor, which takes a shell string

With the IsString instance via OverloadedStrings. If you provide it a string with no spaces (e.g., "date"), it will treat it as a raw command with no arguments (e.g., proc "date" []). If it has spaces, it will use shell.

In all cases, the default for all three streams is to inherit the streams from the parent process. For other settings, see the setters below for default values.

Since: 0.1.0.0

data StreamSpec streamType a

A specification for how to create one of the three standard child streams. See examples below.

Since: 0.1.0.0

data StreamType

Whether a stream is an input stream or output stream. Note that this is from the perspective of the child process, so that a child's standard input stream is an STInput, even though the parent process will be writing to it.

Since: 0.1.0.0

Constructors

STInput

STOutput

data Process stdin stdout stderr

A running process. The three type parameters provide the type of the standard input, standard output, and standard error streams.

Since: 0.1.0.0

ProcessConfig

Smart constructors

proc :: FilePath -> [String] -> ProcessConfig () () ()

Create a ProcessConfig from the given command and arguments.

Since: 0.1.0.0

shell :: String -> ProcessConfig () () ()

Create a ProcessConfig from the given shell command.

Since: 0.1.0.0

Setters

setStdin :: StreamSpec STInput stdin -> ProcessConfig stdin0 stdout stderr -> ProcessConfig stdin stdout stderr

Set the child's standard input stream to the given StreamSpec.

Default: inherit

Since: 0.1.0.0

setStdout :: StreamSpec STOutput stdout -> ProcessConfig stdin stdout0 stderr -> ProcessConfig stdin stdout stderr

Set the child's standard output stream to the given StreamSpec.

Default: inherit

Since: 0.1.0.0

setStderr :: StreamSpec STOutput stderr -> ProcessConfig stdin stdout stderr0 -> ProcessConfig stdin stdout stderr

Set the child's standard error stream to the given StreamSpec.

Default: inherit

Since: 0.1.0.0

setWorkingDir :: FilePath -> ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Set the working directory of the child process.

Default: current process's working directory.

Since: 0.1.0.0

setEnv :: [(String, String)] -> ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Set the environment variables of the child process.

Default: current process's environment.

Since: 0.1.0.0

setCloseFds :: Bool -> ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Should we close all file descriptors besides stdin, stdout, and stderr? See close_fds for more information.

Default: False

Since: 0.1.0.0

setCreateGroup :: Bool -> ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Should we create a new process group?

Default: False

Since: 0.1.0.0

setDelegateCtlc :: Bool -> ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Delegate handling of Ctrl-C to the child. For more information, see delegate_ctlc.

Default: False

Since: 0.1.0.0

setDetachConsole :: Bool -> ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Detach console on Windows, see detach_console.

Default: False

Since: 0.1.0.0

setCreateNewConsole :: Bool -> ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Create new console on Windows, see create_new_console.

Default: False

Since: 0.1.0.0

setNewSession :: Bool -> ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Set a new session with the POSIX setsid syscall, does nothing on non-POSIX. See new_session.

Default: False

Since: 0.1.0.0

setChildGroup :: GroupID -> ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Set the child process's group ID with the POSIX setgid syscall, does nothing on non-POSIX. See child_group.

Default: False

Since: 0.1.0.0

setChildUser :: UserID -> ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Set the child process's user ID with the POSIX setuid syscall, does nothing on non-POSIX. See child_user.

Default: False

Since: 0.1.0.0

setCheckExitCode :: Bool -> ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Should we throw an exception when the process exits with a non-success code?

If set to True, then when stopProcess is called - either directly or via withProcess or other wrappers - the processes exit code will be checked. Any exit code besides ExitSuccess will result in an ExitCodeException being thrown.

Default: False

Since: 0.1.0.0

Stream specs

mkStreamSpec :: StdStream -> (Maybe Handle -> IO (a, IO ())) -> StreamSpec streamType a

Create a new StreamSpec from the given StdStream and a helper function. This function:

Takes as input the raw Maybe Handle returned by the createProcess function. This will be determined by the StdStream argument.

Returns the actual stream value a, as well as a cleanup

function to be run when calling stopProcess.

Since: 0.1.0.0

inherit :: StreamSpec anyStreamType ()

A stream spec which simply inherits the stream of the parent process.

Since: 0.1.0.0

closed :: StreamSpec anyStreamType ()

A stream spec which will close the stream for the child process.

Since: 0.1.0.0

byteStringInput :: ByteString -> StreamSpec STInput ()

An input stream spec which sets the input to the given ByteString. A separate thread will be forked to write the contents to the child process.

Since: 0.1.0.0

byteStringOutput :: StreamSpec STOutput (STM (Either ByteStringOutputException ByteString))

Capture the output of a process in a ByteString.

This function will fork a separate thread to consume all input from the process, and will only make the results available when the underlying Handle is closed. As this is provided as an STM action, you can either check if the result is available, or block until it's ready.

In the event of any exception occurring when reading from the Handle, the result of this function will be a Left value containing a ByteStringOutputException.

Since: 0.1.0.0

createPipe :: StreamSpec anyStreamType Handle

Create a new pipe between this process and the child, and return a Handle to communicate with the child.

Since: 0.1.0.0

useHandleOpen :: Handle -> StreamSpec anyStreamType ()

Use the provided Handle for the child process, and when the process exits, do not close it. This is useful if, for example, you want to have multiple processes write to the same log file sequentially.

Since: 0.1.0.0

useHandleClose :: Handle -> StreamSpec anyStreamType ()

Use the provided Handle for the child process, and when the process exits, close it. If you have no reason to keep the Handle open, you should use this over useHandleOpen.

Since: 0.1.0.0

sink :: MonadIO m => StreamSpec STInput (ConduitM ByteString o m ())

Provide input to a process by writing to a conduit.

Since: 0.1.0.0

source :: MonadIO m => StreamSpec STOutput (ConduitM i ByteString m ())

Read output from a process by read from a conduit.

Since: 0.1.0.0

Launch a process

startProcess :: MonadIO m => ProcessConfig stdin stdout stderr -> m (Process stdin stdout stderr)

Launch a process based on the given ProcessConfig. You should ensure that you close stopProcess on the result. It's usually better to use one of the functions in this module which ensures stopProcess is called, such as withProcess.

Since: 0.1.0.0

stopProcess :: MonadIO m => Process stdin stdout stderr -> m ()

Close a process and release any resources acquired. This will ensure terminateProcess is called, wait for the process to actually exit, and then close out resources allocated for the streams. In the event of any cleanup exceptions being thrown, or if a non-success exit code was received and setCheckExitCode was used, this will throw an exception.

Since: 0.1.0.0

withProcess :: (MonadIO m, MonadMask m) => ProcessConfig stdin stdout stderr -> (Process stdin stdout stderr -> m a) -> m a

Use the bracket pattern to call startProcess and ensure stopProcess is called.

Since: 0.1.0.0

readProcess :: MonadIO m => ProcessConfig stdin stdoutIgnored stderrIgnored -> m (ExitCode, ByteString, ByteString)

Run a process, capture its standard output and error as a ByteString, wait for it to complete, and then return its exit code, output, and error.

Note that any previously used setStdout or setStderr will be overridden.

Since: 0.1.0.0

runProcess :: MonadIO m => ProcessConfig stdin stdout stderr -> m ExitCode

Run the given process, wait for it to exit, and returns its ExitCode.

Since: 0.1.0.0

runProcess_ :: MonadIO m => ProcessConfig stdin stdout stderr -> m ()

Same as runProcess, but ignores the ExitCode.

Since: 0.1.0.0

Interact with a process

Process exit code

waitExitCode :: MonadIO m => Process stdin stdout stderr -> m ExitCode

Wait for the process to exit and then return its ExitCode.

Since: 0.1.0.0

waitExitCodeSTM :: Process stdin stdout stderr -> STM ExitCode

Same as waitExitCode, but in STM.

Since: 0.1.0.0

checkExitCode :: MonadIO m => Process stdin stdout stderr -> m (Maybe ExitCode)

Check if a process has exited and, if so, return its ExitCode.

Since: 0.1.0.0

checkExitCodeSTM :: Process stdin stdout stderr -> STM (Maybe ExitCode)

Same as checkExitCode, but in STM.

Since: 0.1.0.0

Process streams

getStdin :: Process stdin stdout stderr -> stdin

Get the child's standard input stream value.

Since: 0.1.0.0

getStdout :: Process stdin stdout stderr -> stdout

Get the child's standard output stream value.

Since: 0.1.0.0

getStderr :: Process stdin stdout stderr -> stderr

Get the child's standard error stream value.

Since: 0.1.0.0

Exceptions

data ExitCodeException

Exit code generated by stopProcess when setCheckExitCode is True and a process exits with a non-success code. Contains the non-success code, and if any other exceptions occur during cleanup, that exception.

Since: 0.1.0.0

Constructors

ExitCodeException ExitCode (Maybe SomeException)

newtype ByteStringOutputException

Wrapper for when an exception is thrown when reading from a child process, used by byteStringOutput.

Since: 0.1.0.0

Constructors

ByteStringOutputException SomeException