System.Process.Typed

Description

Please see the README.md file for examples of using this API.

Applications using this library MUST use -threaded to link with the threaded version of the RTS. spawnProcess and the others spawn a Haskell thread that blocks until the process exits. Without -threaded this will block all the Haskell threads in your program until the spawned process exits.

Types

data ProcessConfig stdin stdout stderr

An abstract configuration for a process, which can then be launched into an actual running Process. Takes three type parameters, providing the types of standard input, standard output, and standard error, respectively.

There are three ways to construct a value of this type:

With the proc smart constructor, which takes a command name and a list of arguments.

With the shell smart constructor, which takes a shell string

With the IsString instance via OverloadedStrings. If you provide it a string with no spaces (e.g., "date"), it will treat it as a raw command with no arguments (e.g., proc "date" []). If it has spaces, it will use shell.

In all cases, the default for all three streams is to inherit the streams from the parent process. For other settings, see the setters below for default values.

Since: 0.1.0.0

data StreamSpec (streamType :: StreamType) a

A specification for how to create one of the three standard child streams. See examples below.

Since: 0.1.0.0

data StreamType

Whether a stream is an input stream or output stream. Note that this is from the perspective of the child process, so that a child's standard input stream is an STInput, even though the parent process will be writing to it.

Since: 0.1.0.0

Constructors

STInput

STOutput

data Process stdin stdout stderr

A running process. The three type parameters provide the type of the standard input, standard output, and standard error streams.

Since: 0.1.0.0

ProcessConfig

Smart constructors

proc :: FilePath -> [String] -> ProcessConfig () () ()

Create a ProcessConfig from the given command and arguments.

Since: 0.1.0.0

shell :: String -> ProcessConfig () () ()

Create a ProcessConfig from the given shell command.

Since: 0.1.0.0

Setters

setStdin :: StreamSpec STInput stdin -> ProcessConfig stdin0 stdout stderr -> ProcessConfig stdin stdout stderr

Set the child's standard input stream to the given StreamSpec.

Default: inherit

Since: 0.1.0.0

setStdout :: StreamSpec STOutput stdout -> ProcessConfig stdin stdout0 stderr -> ProcessConfig stdin stdout stderr

Set the child's standard output stream to the given StreamSpec.

Default: inherit

Since: 0.1.0.0

setStderr :: StreamSpec STOutput stderr -> ProcessConfig stdin stdout stderr0 -> ProcessConfig stdin stdout stderr

Set the child's standard error stream to the given StreamSpec.

Default: inherit

Since: 0.1.0.0

setWorkingDir :: FilePath -> ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Set the working directory of the child process.

Default: current process's working directory.

Since: 0.1.0.0

setWorkingDirInherit :: ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Inherit the working directory from the parent process.

Since: 0.2.2.0

setEnv :: [(String, String)] -> ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Set the environment variables of the child process.

Default: current process's environment.

Since: 0.1.0.0

setEnvInherit :: ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Inherit the environment variables from the parent process.

Since: 0.2.2.0

setCloseFds :: Bool -> ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Should we close all file descriptors besides stdin, stdout, and stderr? See close_fds for more information.

Default: False

Since: 0.1.0.0

setCreateGroup :: Bool -> ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Should we create a new process group?

Default: False

Since: 0.1.0.0

setDelegateCtlc :: Bool -> ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Delegate handling of Ctrl-C to the child. For more information, see delegate_ctlc.

Default: False

Since: 0.1.0.0

setDetachConsole :: Bool -> ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Detach console on Windows, see detach_console.

Default: False

Since: 0.1.0.0

setCreateNewConsole :: Bool -> ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Create new console on Windows, see create_new_console.

Default: False

Since: 0.1.0.0

setNewSession :: Bool -> ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Set a new session with the POSIX setsid syscall, does nothing on non-POSIX. See new_session.

Default: False

Since: 0.1.0.0

setChildGroup :: GroupID -> ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Set the child process's group ID with the POSIX setgid syscall, does nothing on non-POSIX. See child_group.

Default: False

Since: 0.1.0.0

setChildGroupInherit :: ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Inherit the group from the parent process.

Since: 0.2.2.0

setChildUser :: UserID -> ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Set the child process's user ID with the POSIX setuid syscall, does nothing on non-POSIX. See child_user.

Default: False

Since: 0.1.0.0

setChildUserInherit :: ProcessConfig stdin stdout stderr -> ProcessConfig stdin stdout stderr

Inherit the user from the parent process.

Since: 0.2.2.0

Stream specs

mkStreamSpec :: StdStream -> (ProcessConfig () () () -> Maybe Handle -> IO (a, IO ())) -> StreamSpec streamType a

Create a new StreamSpec from the given StdStream and a helper function. This function:

Takes as input the raw Maybe Handle returned by the createProcess function. This will be determined by the StdStream argument.

Returns the actual stream value a, as well as a cleanup

function to be run when calling stopProcess.

Since: 0.1.0.0

inherit :: StreamSpec anyStreamType ()

A stream spec which simply inherits the stream of the parent process.

Since: 0.1.0.0

closed :: StreamSpec anyStreamType ()

A stream spec which will close the stream for the child process.

Since: 0.1.0.0

byteStringInput :: ByteString -> StreamSpec STInput ()

An input stream spec which sets the input to the given ByteString. A separate thread will be forked to write the contents to the child process.

Since: 0.1.0.0

byteStringOutput :: StreamSpec STOutput (STM ByteString)

Capture the output of a process in a ByteString.

This function will fork a separate thread to consume all input from the process, and will only make the results available when the underlying Handle is closed. As this is provided as an STM action, you can either check if the result is available, or block until it's ready.

In the event of any exception occurring when reading from the Handle, the STM action will throw a ByteStringOutputException.

Since: 0.1.0.0

createPipe :: StreamSpec anyStreamType Handle

Create a new pipe between this process and the child, and return a Handle to communicate with the child.

Since: 0.1.0.0

useHandleOpen :: Handle -> StreamSpec anyStreamType ()

Use the provided Handle for the child process, and when the process exits, do not close it. This is useful if, for example, you want to have multiple processes write to the same log file sequentially.

Since: 0.1.0.0

useHandleClose :: Handle -> StreamSpec anyStreamType ()

Use the provided Handle for the child process, and when the process exits, close it. If you have no reason to keep the Handle open, you should use this over useHandleOpen.

Since: 0.1.0.0

Launch a process

startProcess :: MonadIO m => ProcessConfig stdin stdout stderr -> m (Process stdin stdout stderr)

Launch a process based on the given ProcessConfig. You should ensure that you close stopProcess on the result. It's usually better to use one of the functions in this module which ensures stopProcess is called, such as withProcess.

Since: 0.1.0.0

stopProcess :: MonadIO m => Process stdin stdout stderr -> m ()

Close a process and release any resources acquired. This will ensure terminateProcess is called, wait for the process to actually exit, and then close out resources allocated for the streams. In the event of any cleanup exceptions being thrown this will throw an exception.

Since: 0.1.0.0

withProcess :: ProcessConfig stdin stdout stderr -> (Process stdin stdout stderr -> IO a) -> IO a

Uses the bracket pattern to call startProcess and ensures that stopProcess is called.

In version 0.2.0.0, this function was monomorphized to IO to avoid a dependency on the exceptions package.

Since: 0.1.0.0

withProcess_ :: ProcessConfig stdin stdout stderr -> (Process stdin stdout stderr -> IO a) -> IO a

Same as withProcess, but also calls checkExitCode

In version 0.2.0.0, this function was monomorphized to IO to avoid a dependency on the exceptions package.

Since: 0.1.0.0

readProcess :: MonadIO m => ProcessConfig stdin stdoutIgnored stderrIgnored -> m (ExitCode, ByteString, ByteString)

Run a process, capture its standard output and error as a ByteString, wait for it to complete, and then return its exit code, output, and error.

Note that any previously used setStdout or setStderr will be overridden.

Since: 0.1.0.0

readProcess_ :: MonadIO m => ProcessConfig stdin stdoutIgnored stderrIgnored -> m (ByteString, ByteString)

Same as readProcess, but instead of returning the ExitCode, checks it with checkExitCode.

Since: 0.1.0.0

runProcess :: MonadIO m => ProcessConfig stdin stdout stderr -> m ExitCode

Run the given process, wait for it to exit, and returns its ExitCode.

Since: 0.1.0.0

runProcess_ :: MonadIO m => ProcessConfig stdin stdout stderr -> m ()

Same as runProcess, but instead of returning the ExitCode, checks it with checkExitCode.

Since: 0.1.0.0

readProcessStdout :: MonadIO m => ProcessConfig stdin stdoutIgnored stderr -> m (ExitCode, ByteString)

Same as readProcess, but only read the stdout of the process. Original settings for stderr remain.

Since: 0.2.1.0

readProcessStdout_ :: MonadIO m => ProcessConfig stdin stdoutIgnored stderr -> m ByteString

Same as readProcessStdout, but instead of returning the ExitCode, checks it with checkExitCode.

Since: 0.2.1.0

readProcessStderr :: MonadIO m => ProcessConfig stdin stderrIgnored stderr -> m (ExitCode, ByteString)

Same as readProcess, but only read the stderr of the process. Original settings for stderr remain.

Since: 0.2.1.0

readProcessStderr_ :: MonadIO m => ProcessConfig stdin stderrIgnored stderr -> m ByteString

Same as readProcessStderr, but instead of returning the ExitCode, checks it with checkExitCode.

Since: 0.2.1.0

Interact with a process

Process exit code

waitExitCode :: MonadIO m => Process stdin stdout stderr -> m ExitCode

Wait for the process to exit and then return its ExitCode.

Since: 0.1.0.0

waitExitCodeSTM :: Process stdin stdout stderr -> STM ExitCode

Same as waitExitCode, but in STM.

Since: 0.1.0.0

getExitCode :: MonadIO m => Process stdin stdout stderr -> m (Maybe ExitCode)

Check if a process has exited and, if so, return its ExitCode.

Since: 0.1.0.0

getExitCodeSTM :: Process stdin stdout stderr -> STM (Maybe ExitCode)

Same as getExitCode, but in STM.

Since: 0.1.0.0

checkExitCode :: MonadIO m => Process stdin stdout stderr -> m ()

Wait for a process to exit, and ensure that it exited successfully. If not, throws an ExitCodeException.

Since: 0.1.0.0

checkExitCodeSTM :: Process stdin stdout stderr -> STM ()

Same as checkExitCode, but in STM.

Since: 0.1.0.0

Process streams

getStdin :: Process stdin stdout stderr -> stdin

Get the child's standard input stream value.

Since: 0.1.0.0

getStdout :: Process stdin stdout stderr -> stdout

Get the child's standard output stream value.

Since: 0.1.0.0

getStderr :: Process stdin stdout stderr -> stderr

Get the child's standard error stream value.

Since: 0.1.0.0

Exceptions

data ExitCodeException

Exception thrown by checkExitCode in the event of a non-success exit code. Note that checkExitCode is called by other functions as well, like runProcess_ or readProcess_.

Since: 0.1.0.0

Constructors

ExitCodeException

Fields

eceExitCode :: ExitCode

eceProcessConfig :: ProcessConfig () () ()

eceStdout :: ByteString

eceStderr :: ByteString

data ByteStringOutputException

Wrapper for when an exception is thrown when reading from a child process, used by byteStringOutput.

Since: 0.1.0.0

Constructors

ByteStringOutputException SomeException (ProcessConfig () () ())

Unsafe functions

unsafeProcessHandle :: Process stdin stdout stderr -> ProcessHandle

Take ProcessHandle out of the Process. This method is needed in cases one need to use low level functions from the process package. Use cases for this method are:

Send a special signal to the process.

Terminate the process group instead of terminating single process.

Use platform specific API on the underlying process.

This method is considered unsafe because the actions it performs on the underlying process may overlap with the functionality that typed-process provides. For example the user should not call waitForProcess on the process handle as eiter waitForProcess or stopProcess will lock. Additionally, even if process was terminated by the terminateProcess or by sending signal, stopProcess should be called either way in order to cleanup resources allocated by the typed-process.

Since: 0.1.1